+++ my_enume.rb
@@ -1,0 +1,17 @@
+module MyEnumerable
+  def all?()
+    each { |item| return false unless yield item }
+    true
+  end
+
+  def any?()
+    each { |item| return true if yield item }
+    false
+  end
+
+  def filter?()
+    array = []
+    each { |item| array << item if yield item }
+    array
+  end
+end
