+++ my_list.rb
@@ -1,0 +1,13 @@
+require_relative './my_enume'
+
+class MyList
+  include MyEnumerable
+
+  def initialize(*list)
+    @list = list
+  end
+
+  def each(&block)
+    @list.each(&block)
+  end
+end
